requirements.txt
openai>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
streamlit
gTTS
pydub
pyttsx3
streamlit-webrtc
sounddevice
numpy
python-dotenv==1.0.0 